requirements-dev.txt
# Development dependencies for Antman project

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0

# Code quality
ruff==0.1.6
black==23.11.0
isort==5.12.0
pre-commit==3.5.0

# Security scanning
safety==2.3.5
bandit==1.7.5
pip-audit==2.6.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2
ipdb==0.13.13

# Testing utilities
coverage==7.3.2
model-bakery==1.17.0
responses==0.24.1

# YAML parsing for CI tests
PyYAML==6.0.1
